Weather-app
Demonstrate the Asynchronous executions of instructions using node setTimeOut()
That method takes two arguments. The First is a callback function, and the second is a delay that set time after which the function will be called.
This setTimeout() does not block the flow of execution, while setTimeout is waiting for the delay to expire, the rest of the instructions of the program is executing.